# packages/server/docs/swagger.json
{
  "info": {
    "title": "Sonar Server HTTP API",
    "version": "0.1.0",
    "license": {
      "name": "GPL-3.0"
    },
    "description": "[Specification JSON](/api-spec)"
  },
  "paths": {
    "/api/login": {
      "post": {
        "summary": "/api/login",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/create-access-code": {
      "post": {
        "summary": "/api/create-access-code",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/info": {
      "get": {
        "summary": "/api/info",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/commands/.websocket": {
      "get": {
        "summary": "/api/commands/.websocket",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/collection": {
      "post": {
        "summary": "/api/collection",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/collection/{collection}": {
      "patch": {
        "summary": "/api/collection/{collection}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "get": {
        "summary": "/api/collection/{collection}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/fs/{drive}/*": {
      "get": {
        "summary": "/api/collection/{collection}/fs/{drive}/*",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "put": {
        "summary": "/api/collection/{collection}/fs/{drive}/*",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/fs/{drive}": {
      "get": {
        "summary": "/api/collection/{collection}/fs/{drive}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/db": {
      "put": {
        "summary": "/api/collection/{collection}/db",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/db/{id}": {
      "put": {
        "summary": "/api/collection/{collection}/db/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "delete": {
        "summary": "/api/collection/{collection}/db/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/db/{key}/{seq}": {
      "get": {
        "summary": "/api/collection/{collection}/db/{key}/{seq}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "key",
            "in": "path",
            "required": true
          },
          {
            "name": "seq",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/sync": {
      "get": {
        "summary": "/api/collection/{collection}/sync",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/query/{name}": {
      "post": {
        "summary": "/api/collection/{collection}/query/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/schema": {
      "get": {
        "summary": "/api/collection/{collection}/schema",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "post": {
        "summary": "/api/collection/{collection}/schema",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/feed/{key}": {
      "put": {
        "summary": "/api/collection/{collection}/feed/{key}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "key",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/debug": {
      "get": {
        "summary": "/api/collection/{collection}/debug",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/subscription/{name}": {
      "put": {
        "summary": "/api/collection/{collection}/subscription/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "get": {
        "summary": "/api/collection/{collection}/subscription/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/subscription/{name}/sse": {
      "get": {
        "summary": "/api/collection/{collection}/subscription/{name}/sse",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/subscription/{name}/{cursor}": {
      "post": {
        "summary": "/api/collection/{collection}/subscription/{name}/{cursor}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "name": "cursor",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/events": {
      "get": {
        "summary": "/api/collection/{collection}/events",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/reindex": {
      "post": {
        "summary": "/api/collection/{collection}/reindex",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/collection/{collection}/fs-info": {
      "get": {
        "summary": "/api/collection/{collection}/fs-info",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/login": {
      "post": {
        "summary": "/api/v1/login",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/v1/create-access-code": {
      "post": {
        "summary": "/api/v1/create-access-code",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/v1/info": {
      "get": {
        "summary": "/api/v1/info",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/v1/commands/.websocket": {
      "get": {
        "summary": "/api/v1/commands/.websocket",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/v1/collection": {
      "post": {
        "summary": "/api/v1/collection",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}": {
      "patch": {
        "summary": "/api/v1/collection/{collection}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "get": {
        "summary": "/api/v1/collection/{collection}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/fs/{drive}/*": {
      "get": {
        "summary": "/api/v1/collection/{collection}/fs/{drive}/*",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "put": {
        "summary": "/api/v1/collection/{collection}/fs/{drive}/*",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/fs/{drive}": {
      "get": {
        "summary": "/api/v1/collection/{collection}/fs/{drive}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "drive",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/db": {
      "put": {
        "summary": "/api/v1/collection/{collection}/db",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/db/{id}": {
      "put": {
        "summary": "/api/v1/collection/{collection}/db/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "delete": {
        "summary": "/api/v1/collection/{collection}/db/{id}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/db/{key}/{seq}": {
      "get": {
        "summary": "/api/v1/collection/{collection}/db/{key}/{seq}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "key",
            "in": "path",
            "required": true
          },
          {
            "name": "seq",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/sync": {
      "get": {
        "summary": "/api/v1/collection/{collection}/sync",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/query/{name}": {
      "post": {
        "summary": "/api/v1/collection/{collection}/query/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/schema": {
      "get": {
        "summary": "/api/v1/collection/{collection}/schema",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "post": {
        "summary": "/api/v1/collection/{collection}/schema",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/feed/{key}": {
      "put": {
        "summary": "/api/v1/collection/{collection}/feed/{key}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "key",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/debug": {
      "get": {
        "summary": "/api/v1/collection/{collection}/debug",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/subscription/{name}": {
      "put": {
        "summary": "/api/v1/collection/{collection}/subscription/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      },
      "get": {
        "summary": "/api/v1/collection/{collection}/subscription/{name}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/subscription/{name}/sse": {
      "get": {
        "summary": "/api/v1/collection/{collection}/subscription/{name}/sse",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/subscription/{name}/{cursor}": {
      "post": {
        "summary": "/api/v1/collection/{collection}/subscription/{name}/{cursor}",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          },
          {
            "name": "name",
            "in": "path",
            "required": true
          },
          {
            "name": "cursor",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/events": {
      "get": {
        "summary": "/api/v1/collection/{collection}/events",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/reindex": {
      "post": {
        "summary": "/api/v1/collection/{collection}/reindex",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    },
    "/api/v1/collection/{collection}/fs-info": {
      "get": {
        "summary": "/api/v1/collection/{collection}/fs-info",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true
          }
        ],
        "responses": {}
      }
    }
  },
  "swagger": "2.0"
}